chore: cleanup, doc
