+++ pyeudiw/openid4vp/interface.py
@@ -8,14 +8,30 @@ from pyeudiw.openid4vp.schemas.response import AuthorizeResponsePayload
 
 class AuthorizationResponseParser:
     """
-    DirectPostParser is an interface intended to parse direct POST http responses
+    AuthorizationResponseParser is an interface intended to parse direct POST
+    http responses.
+
+    An authorization parser is meant to just parse and eventually validate the
+    "lower" applicaiton layer of the transmission, that is, it is used to
+    extract an authorization response from the HTTP layer.
+    It SHOULD NOT be used to validate the actual content of the response, that
+    is, it SHOULD NOT try to validate vp_tokens, presentation_submissions, etc.
+    This is a delicate task that that is best suited for a different, dedicated
+    object, method or interface.
     """
 
     def parse_and_validate(self, context: satosa.context.Context) -> AuthorizeResponsePayload:
         """
-        Parse (and optionally validate) a satosa http request, wrapped in its own context.
+        Parse (and optionally validate) a satosa http request, wrapped in its own
+        context, in order to extract an auhtorization response.
         The validation step might include verification tasks; for example if the data is
         reepresented as a jwt, the validation should perform a check on the jwt validity.
+
+        The concrete implementation SHOULD NOT be used to validate the actual content
+        of the response, that is, it SHOULD NOT try to validate vp_tokens,
+        presentation_submissions, etc.
+        This is a delicate task that that is best suited for a different, dedicated
+        object, method or interface.
 
         :param context: an http request wrapped in its own satosa context
         :type context: satosa.context.Context
